11_dars/requirements.txt
aerich==0.7.2
aiosqlite==0.17.0
annotated-types==0.7.0
anyio==4.4.0
click==8.1.7
dictdiffer==0.9.0
fastapi==0.114.0
h11==0.14.0
idna==3.8
iso8601==1.1.0
pydantic==2.9.1
pydantic_core==2.23.3
pypika-tortoise==0.1.6
python-multipart==0.0.9
pytz==2024.1
sniffio==1.3.1
starlette==0.38.5
tomlkit==0.13.2
tortoise-orm==0.21.6
typing_extensions==4.12.2
uvicorn==0.30.6
